refactor: Migrate test code from unittest to pytest style (#4)
- Remove unused tests.py files
- Move test code from a_user to a_apis directory
- Convert test cases to pytest format
